+++ src/hooks/index.ts
@@ -1,1 +1,4 @@
 export * from "./useHeaderData";
+export * from "./useRecentBlogPosts";
+export * from "./useRecentAnnouncements";
+export * from "./useDefaultHeroImage";
