+++ src/sections/RecentAnnouncements.tsx
@@ -1,6 +1,6 @@
 import React from "react";
-import { graphql, useStaticQuery } from "gatsby";
 import { RecentAnnouncement } from "components";
+import { useRecentAnnouncements } from "hooks";
 import { SectionBase } from "types";
 import { Section } from "components";
 
@@ -8,39 +8,8 @@ export interface RecentAnnouncementsSectionData extends SectionBase {
   type: "recentAnnouncements";
 }
 
-type AnnouncementNode = {
-  frontmatter: {
-    title: string;
-    summary: string;
-    date: string;
-  };
-  fields: {
-    slug: string;
-  };
-};
-
 export const RecentAnnouncements: React.FC<RecentAnnouncementsSectionData> = () => {
-  const data = useStaticQuery(graphql`
-    query {
-      allMarkdownRemark(
-        filter: { frontmatter: { type: { eq: "announcements" } } }
-        sort: { fields: frontmatter___date, order: DESC }
-        limit: 2
-      ) {
-        nodes {
-          frontmatter {
-            title
-            date(formatString: "MMM DD, YYYY")
-          }
-          fields {
-            slug
-          }
-        }
-      }
-    }
-  `);
-
-  const announcements: AnnouncementNode[] = data.allMarkdownRemark.nodes;
+  const announcements = useRecentAnnouncements();
 
   return (
     <Section>
