+++ src/styles/components/header.scss
@@ -172,7 +172,12 @@
     cursor: pointer;
     font-size: 20px;
     padding: 0 22px;
+    display: none;
     color: color($theme-color-base);
+
+    @include at-media(desktop) {
+      display: block;
+    }
   }
 
   .dds-search-form-wrapper {
